Updated and added files
Added the rest of the blinky files for ease of use.

Updated test.c code which has the basic structure and functions for the code.
Updated main.c (Edited Blinky main) for simple hardware testing

I took some of the functions from test.c and put them in main.c just for hardware testing purposes. This main.c is still basically the Blinky main.c but changed. (Though most of what is in the blinky program is in the test.c in terms of initialization and such.